Fix handle force curves being shifted by one
Refactor torque handling by adding torqueBeforeFlank in StrokeService to
shift the included force values by one data point so the right values in
the force curve are shown. When drive or recovery is detected the first
and last drive force the data before the datapoint that is currently
being processed.