Deprecate `minimumRecoverySlopeMargin` setting
Remove minimumRecoverySlopeMargin from settings as it is obsolete and
the driveHandleForcesMaxCapacity setting makes it redundant.

Update the Stroke Detection Settings BLE characteristic payload as well
as update the settings write control point. Update documentation to
reflect new BLE schema and deprecation.

This is a breaking change as the Stroke Detection Settings BLE payload
was changed from 15 bytes to 11 bytes; old clients sending the 15-byte
payload will be rejected and must be updated.